test: fix RuntimeWarnings from improper AsyncMock usage in tests (#1344)
* test: fix RuntimeWarnings from improper AsyncMock usage in tests

Fixed RuntimeWarnings caused by using AsyncMock() for synchronous methods
across multiple test files. The issue occurred when tests mocked sync methods
(like db.commit(), response.raise_for_status(), Redis.from_url()) with
AsyncMock, creating unawaited coroutines.

Changes:
- Use Mock()/MagicMock() for synchronous methods instead of AsyncMock()
- Fixed SQLAlchemy session methods (commit, refresh, close)
- Fixed HTTP response methods (raise_for_status)
- Fixed Redis client methods (from_url, pubsub, close)
- Fixed asyncio methods (create_task return values)
- Fixed event loop methods (call_soon_threadsafe)
- Removed unused event_loop fixture from conftest.py

Files modified:
- tests/unit/mcpgateway/test_reverse_proxy.py: Fixed process/task mocks
- tests/unit/mcpgateway/cache/test_session_registry.py: Fixed db session mocks
- tests/unit/mcpgateway/cache/test_session_registry_extended.py: Fixed Redis mocks
- tests/unit/mcpgateway/services/test_tool_service.py: Fixed db/HTTP mocks
- tests/unit/mcpgateway/services/test_logging_service_comprehensive.py: Fixed event loop mocks
- tests/unit/mcpgateway/services/test_role_service.py: Fixed patch.object usage
- tests/conftest.py: Removed deprecated event_loop fixture

Test results:
- 3108 tests passed
- Reduced RuntimeWarnings from many to 4 false-positive attributions
- All individual tests pass with no RuntimeWarnings


* chore: remove accidentally committed .orig file

Remove test_translate.py.orig which was accidentally included in the
previous commit. This is a merge/rebase artifact that should not be
tracked in version control.


---------

+++ mcpgateway/reverse_proxy.py
@@ -258,7 +258,7 @@ class ReverseProxyClient:
         token: Optional[str] = None,
         reconnect_delay: float = DEFAULT_RECONNECT_DELAY,
         max_retries: int = DEFAULT_MAX_RETRIES,
-        keepalive_interval: int = DEFAULT_KEEPALIVE_INTERVAL,
+        keepalive_interval: float = DEFAULT_KEEPALIVE_INTERVAL,
     ):
         """Initialize reverse proxy client.
